Moved classes into folders.
