+++ smtp-config.css
@@ -2,35 +2,35 @@
 	display: inline-block;
 	padding: .25em .4em;
 	padding-right: .6em;
-    padding-left: .6em;
-    font-size: 75%;
-    font-weight: 700;
-    line-height: 1;
-    text-align: center;
-    white-space: nowrap;
-    vertical-align: baseline;
+	padding-left: .6em;
+	font-size: 75%;
+	font-weight: 700;
+	line-height: 1;
+	text-align: center;
+	white-space: nowrap;
+	vertical-align: baseline;
 	border-radius: 10rem;
-	
+
 	color: #fff;
-    background-color: #6c757d;
+	background-color: #6c757d;
 }
 
 .wpsmtp-badge.wpsmtp-badge-info {
 	color: #fff;
-    background-color: #17a2b8;
+	background-color: #17a2b8;
 }
 
 .wpsmtp-badge.wpsmtp-badge-success {
 	color: #fff;
-    background-color: #28a745;
+	background-color: #28a745;
 }
 
 .wpsmtp-badge.wpsmtp-badge-warning {
 	color: #212529;
-    background-color: #ffc107;
+	background-color: #ffc107;
 }
 
 .wpsmtp-badge.wpsmtp-badge-danger {
-    color: #fff;
-    background-color: #dc3545;
+	color: #fff;
+	background-color: #dc3545;
 }
