+++ src/log/class-log.php
