+++ src/log/class-logtable.php
