+++ src/mail/class-mail.php
