+++ src/settings/class-settings.php
@@ -12,6 +12,8 @@ namespace wpsimplesmtp;
 use wpsimplesmtp\Options;
 use wpsimplesmtp\Log;
 use wpsimplesmtp\LogTable;
+use wpsimplesmtp\Mailtest;
+use wpsimplesmtp\MailView;
 
 /**
  * Handles the visibility and setup with the WordPress Settings API.
@@ -45,12 +47,13 @@ class Settings {
 		add_action( 'admin_menu', [ &$this, 'add_admin_menu' ] );
 		add_action( 'admin_init', [ &$this, 'settings_init' ] );
 		add_action( 'admin_init', [ &$this, 'settings_test_init' ] );
-		add_action( 'admin_post_ss_test_email', [ &$this, 'test_email_handler' ] );
 		add_filter( 'pre_update_option_wpssmtp_smtp', [ &$this, 'post_processing' ] );
 
 		$this->options   = new Options();
 		$this->log       = new Log();
 		$this->log_table = new LogTable();
+		$this->mail_test = new Mailtest();
+		$this->mail_view = new MailView();
 	}
 
 	/**
@@ -66,7 +69,7 @@ class Settings {
 		$return = false;
 		if ( isset( $_REQUEST['ssnonce'], $_REQUEST['eid'], $_REQUEST['resend'] ) && wp_verify_nonce( sanitize_key( $_REQUEST['ssnonce'] ), 'wpss_action' ) ) {
 			$return = true;
-			$resp   = $this->resend_email( intval( $_REQUEST['eid'] ) );
+			$resp   = $this->mail_test->resend_email( intval( $_REQUEST['eid'] ) );
 			if ( $resp ) {
 				?>
 				<div class="notice notice-success is-dismissible">
@@ -101,7 +104,7 @@ class Settings {
 		}
 
 		if ( isset( $_REQUEST['eid'] ) && ! $return ) {
-			$this->render_email_view( intval( $_REQUEST['eid'] ) );
+			$this->mail_view->render_email_view( intval( $_REQUEST['eid'] ) );
 		} else {
 			$this->render_settings();
 		}
@@ -191,87 +194,6 @@ class Settings {
 			'wpsimplesmtp_smtp_test',
 			'wpsimplesmtp_test_email'
 		);
-	}
-
-	/**
-	 * Custom admin endpoint to dispatch a test email.
-	 */
-	public function test_email_handler() {
-		if ( isset( $_REQUEST['_wpnonce'], $_REQUEST['_wp_http_referer'], $_REQUEST['wpssmtp_test_email_recipient'] ) && wp_verify_nonce( sanitize_key( $_REQUEST['_wpnonce'] ), 'simple-smtp-test-email' ) ) {
-			$is_html      = ( isset( $_REQUEST['wpssmtp_test_email_is_html'] ) ) ? true : false;
-			$content_type = ( $is_html ) ? 'Content-Type: text/html' : 'Content-Type: text/plain';
-			$content      = __( 'This email proves that your settings are correct.', 'simple-smtp' ) . PHP_EOL . get_bloginfo( 'url' );
-
-			if ( $is_html ) {
-				$html_email  = '<body>';
-				$html_email .= '<div style="text-align: center;margin-top: 5%;font-size: 4em;">' . __( '&#9989;', 'simple-smtp' ) . '</div>';
-				$html_email .= '<h1 style="font-family: sans-serif;text-align: center;font-size: 4em;">' . __( 'This is a test email', 'simple-smtp' ) . '</h1>';
-				$html_email .= '<p style="font-family: sans-serif;text-align: center;font-size: 1em;">' . $content . '</p>';
-				$html_email .= '</body>';
-
-				$content = wp_kses_post( $html_email );
-			}
-
-			// Sanitize rule disabled here as it doesn't detect the later sanitize call. Feel free to refactor.
-			// phpcs:disable WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
-			$recipients = explode( ';', wp_unslash( $_REQUEST['wpssmtp_test_email_recipient'] ) );
-			$recp_count = count( $recipients );
-			// phpcs:enable
-			for ( $i = 0; $i < $recp_count; $i++ ) {
-				$recipients[ $i ] = sanitize_email( trim( $recipients[ $i ] ) );
-			}
-
-			wp_mail(
-				$recipients,
-				// translators: %s is the website name.
-				sprintf( __( 'Test email from %s', 'simple-smtp' ), get_bloginfo( 'name' ) ),
-				$content,
-				[ 'x-test: WP SMTP', $content_type ],
-			);
-
-			wp_safe_redirect( admin_url( 'options-general.php?page=wpsimplesmtp' ) );
-			exit;
-		} else {
-			wp_die( esc_attr_e( 'You are not permitted to send a test email.', 'simple-smtp' ) );
-		}
-	}
-
-	/**
-	 * Resends an email.
-	 *
-	 * @param integer $email_id Email/log ID to resend.
-	 * @return boolean
-	 */
-	public function resend_email( $email_id ) {
-		$email       = $this->log->get_log_entry_by_id( $email_id );
-		$attachments = $this->log->get_log_entry_attachments( $email_id );
-		$recipients  = implode( ', ', json_decode( get_post_meta( $email->ID, 'recipients', true ) ) );
-		$headers     = json_decode( get_post_meta( $email->ID, 'headers', true ) );
-		$opts        = get_option( 'wpss_resent', [] );
-
-		$attachpaths = [];
-		foreach ( $attachments as $attachment ) {
-			if ( $attachment->exists() ) {
-				$attachpaths[] = $attachment->file_path();
-			}
-		}
-
-		if ( isset( $email ) && ! in_array( $email_id, $opts, true ) ) {
-			$opts[] = $email_id;
-			update_option( 'wpss_resent', $opts );
-
-			wp_mail(
-				$recipients,
-				$email->post_title,
-				$email->post_content,
-				$headers,
-				$attachpaths
-			);
-
-			return true;
-		} else {
-			return false;
-		}
 	}
 
 	/**
@@ -414,92 +336,6 @@ class Settings {
 			</form>
 		</div>
 		<?php
-	}
-
-	/**
-	 * Render the email with useful information.
-	 *
-	 * @param integer $id Email log ID.
-	 * @return void Prints to page.
-	 */
-	private function render_email_view( $id ) {
-		$log         = $this->log->get_log_entry_by_id( $id );
-		$attachments = $this->log->get_log_entry_attachments( $id );
-		$recset      = ( in_array( (int) $id, get_option( 'wpss_resent', [] ), true ) ) ? ' disabled' : '';
-		$resend_url  = add_query_arg(
-			[
-				'eid'     => $id,
-				'ssnonce' => wp_create_nonce( 'wpss_action' ),
-			],
-			menu_page_url( 'wpsimplesmtp', false )
-		) . '&resend';
-
-		if ( current_user_can( 'administrator' ) && isset( $log ) ) {
-			$recipients = implode( ', ', json_decode( get_post_meta( $log->ID, 'recipients', true ) ) );
-			$date       = gmdate( get_option( 'time_format' ) . ', ' . get_option( 'date_format' ), strtotime( get_post_meta( $log->ID, 'timestamp', true ) ) );
-
-			$content = '';
-			if ( isset( $log->headers ) && false !== strpos( $log->headers, 'Content-Type: text\/html' ) ) {
-				$content = wp_kses_post( $log->post_content );
-			} else {
-				$content = wp_kses_post( '<pre>' . $log->post_content . '</pre>' );
-			}
-			?>
-			<div class="wrap">
-				<h1><?php esc_html_e( 'View Email', 'simple-smtp' ); ?></h1>
-				<div id="poststuff">
-					<div id="post-body" class="metabox-holder columns-2">
-						<div id="post-body-content">
-							<div class="postbox">
-								<h2 class="hndle"><?php echo esc_html( $log->post_title ); ?></h2>			
-								<div class="inside">
-									<?php echo wp_kses_post( $content ); ?>
-								</div>	
-							</div>
-						</div>
-						<div id="postbox-container-1" class="postbox-container">
-							<div class="stuffbox">
-								<h2 class="hndle"><?php esc_html_e( 'Information', 'simple-smtp' ); ?></h2>
-								<div class="inside">
-									<div id="minor-publishing">
-										<div id="misc-publishing-actions">
-											<div class="misc-pub-section"><?php esc_html_e( 'Recipient(s)', 'simple-smtp' ); ?>: <strong><?php echo esc_html( $recipients ); ?></strong></div>
-											<div class="misc-pub-section"><?php esc_html_e( 'Date sent', 'simple-smtp' ); ?>: <strong><?php echo esc_html( $date ); ?></strong></div>
-											<?php if ( ! empty( $attachments ) ) : ?>
-												<div class="misc-pub-section">
-													<?php esc_html_e( 'Attachment(s)', 'simple-smtp' ); ?>:
-													<ol>
-														<?php foreach ( $attachments as $attachment ) : ?>
-															<li>
-																<?php echo esc_html( $attachment->basename() ); ?>
-																<?php if ( ! $attachment->exists() ) : ?>
-																	<span class="wpsmtp-badge wpsmtp-badge-warning"><?php esc_html_e( 'File missing', 'simple-smtp' ); ?></span>
-																<?php endif; ?>
-															</li>
-														<?php endforeach; ?>
-													</ol>
-												</div>
-											<?php endif; ?>
-										</div>
-										<div class="clear"></div>
-									</div>
-									<div id="major-publishing-actions">
-										<div id="publishing-action">
-											<a href="<?php echo esc_html( $resend_url ); ?>" class="button button-primary button-large <?php echo esc_attr( $recset ); ?>"><?php esc_html_e( 'Resend', 'simple-smtp' ); ?></a>
-										</div>
-										<div class="clear"></div>
-									</div>
-								</div>
-							</div>
-						</div>
-					</div>
-					<br class="clear">
-				</div>
-			</div>
-			<?php
-		} else {
-			wp_die( 'No email found.' );
-		}
 	}
 
 	/**
